split into https-redirector-client and https-redirector-server, converted to typescript, no logic changes so far
