--- API.mjs
@@ -1,3 +1,0 @@
-export { default as Hub } from './src/Hub.mjs';
-export { default as WebServer } from './src/WebServer.mjs';
-export { default as HubUplinkClient } from './src/HubUplinkClient.mjs';
