--- https-reflector-client.mjs
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-
-import main from './src/uplink-main.mjs';
-main();
