+++ https-reflector-client/https-reflector-client.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+const main = require('./build/main.js');
+main.default();
