+++ https-reflector-client/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "https-reflector-client",
+  "version": "2.0.0",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ferguson/https-reflector.git"
+  },
+  "author": "Jon Ferguson <jon@jef.us>",
+  "description": "HTTPS Reflector Client",
+  "license": "MIT",
+  "private": true,
+  "main": "build/API.js",
+  "scripts": {
+    "build": "tsc 2>&1 | egrep -v '^node_modules/(socket[.]io|@socket[.]io|socket[.]io-client|engine[.]io-client)/'"
+  },
+  "dependencies": {
+    "commander": "^7",
+    "pump": "^3.0.0",
+    "socket.io-client": "^4.7.5",
+    "websocket-stream": "^5.5.2",
+    "ws": "^8.17.1"
+  },
+  "devDependencies": {
+    "@types/node": "6.0.94",
+    "nodemon": "^3.1.3",
+    "typescript": "2.5.3"
+  }
+}
