+++ https-reflector-server/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "https-reflector-server",
+  "version": "2.0.0",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ferguson/https-reflector.git"
+  },
+  "author": "Jon Ferguson <jon@jef.us>",
+  "description": "HTTPS Reflector Server",
+  "license": "MIT",
+  "private": true,
+  "main": "build/API.js",
+  "scripts": {
+    "build": "tsc 2>&1 | egrep -v '^node_modules/(socket[.]io|[@]socket[.]io|socket[.]io-client|engine[.]io-client)/'"
+  },
+  "dependencies": {
+    "commander": "^7",
+    "express": "^4.18.2",
+    "express-async-handler": "^1.2.0",
+    "node-fetch": "^2",
+    "pump": "^3.0.0",
+    "serve-favicon": "^2.5.0",
+    "snake-camel": "^1.0.9",
+    "socket.io": "4.1.3",
+    "vhost": "^3.0.2",
+    "websocket-stream": "^5.5.2",
+    "ws": "^8.17.1"
+  },
+  "devDependencies": {
+    "@types/node": "6.0.94",
+    "nodemon": "^3.1.3",
+    "typescript": "2.5.3"
+  }
+}
