+++ https-reflector-server/run.sh
@@ -6,5 +6,5 @@ source ./activate
 
 export HTTPS_REFLECTOR_PUBLIC_STATIC_DIR=/var/www/some-https-reflector-server.org
 
-echo "starting https-reflector"
-exec node --experimental-modules ./https-reflector.mjs --use-https --redirect-http --hostname="*.some-https-reflector-server.org"
+echo "starting https-reflector-server"
+exec node ./https-reflector-server.js --use-https --redirect-http --hostname="*.some-https-reflector-server.org"
