--- https-reflector.mjs
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-
-import main from './src/main.mjs';
-main();
