+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "https-reflector",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ferguson/https-reflector.git"
@@ -9,25 +9,18 @@
   "description": "HTTPS Reflector",
   "license": "MIT",
   "private": true,
-  "main": "./API.mjs",
   "scripts": {
-    "start": "node https-reflector.js"
+    "build": "yarn build-client; yarn build-server",
+    "build-client": "(cd https-reflector-client; yarn build)",
+    "build-server": "(cd https-reflector-server; yarn build)"
   },
-  "dependencies": {
-    "commander": "^7",
-    "express": "^4.18.2",
-    "express-async-handler": "^1.2.0",
-    "node-fetch": "^2",
-    "pump": "^3.0.0",
-    "serve-favicon": "^2.5.0",
-    "snake-camel": "^1.0.9",
-    "socket.io": "4.1.3",
-    "socket.io-client": "^4.7.5",
-    "vhost": "^3.0.2",
-    "websocket-stream": "^5.5.2",
-    "ws": "^8.17.1"
-  },
-  "devDependencies": {
-    "nodemon": "^3.1.3"
+  "workspaces": {
+    "packages": [
+      "https-reflector-client",
+      "https-reflector-server"
+    ],
+    "nohoist": [
+      "**"
+    ]
   }
 }
